ShvPath::matchWild() test
